Refactor input parser into Parser class
