+++ src/main/java/task/ToDo.java
@@ -22,7 +22,7 @@ public class ToDo extends Task {
     }
 
     public static Task fromJsonString(String jsonString) throws InvalidTaskException {
-        Map<String, String> arguments = parseJsonString(jsonString);
+        Map<String, String> arguments = Parser.parseJsonString(jsonString);
         String inputLine = String.format("toDo %s", arguments.get("description"));
         ToDo toDo = new ToDo(inputLine);
         toDo.isCompleted = arguments.get("isCompleted").compareTo("true") == 0;
